# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book Api documentation",
        "description": "Book Api documentation",
        "contact": {
            "email": "kuriyamehul@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/v1/rentals": {
            "post": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Create a rental",
                "description": "Rent a book for a user.",
                "operationId": "74b4033eeb2ec5d519c682808a80b01e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "book_id": {
                                        "description": "ID of the book to rent",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "description": "ID of the user renting the book",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rental created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID of the rental",
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "description": "ID of the user",
                                            "type": "integer"
                                        },
                                        "book_id": {
                                            "description": "ID of the rented book",
                                            "type": "integer"
                                        },
                                        "rented_at": {
                                            "description": "Date and time when the book was rented",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "due_at": {
                                            "description": "Due date for returning the book",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "This book is already rented."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/rentals/return/{id}": {
            "put": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Return a rented book",
                "description": "Mark a rented book as returned.",
                "operationId": "43e221d193680f5595aef0b6146a3ac6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the rental to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rental returned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID of the rental",
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "description": "ID of the user",
                                            "type": "integer"
                                        },
                                        "book_id": {
                                            "description": "ID of the rented book",
                                            "type": "integer"
                                        },
                                        "rented_at": {
                                            "description": "Date and time when the book was rented",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "due_at": {
                                            "description": "Due date for returning the book",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "returned_at": {
                                            "description": "Date and time when the book was returned",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "overdue": {
                                            "description": "Indicates if the rental is overdue",
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "This rental has already been returned."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/rentals/history": {
            "get": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Get rental history",
                "description": "Retrieve the rental history for a specific user.",
                "operationId": "4df628d78cb2ca0df5dfc0494c3d7551",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID of the user to get rental history for",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rental history retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID of the rental",
                                                "type": "integer"
                                            },
                                            "user_id": {
                                                "description": "ID of the user",
                                                "type": "integer"
                                            },
                                            "book_id": {
                                                "description": "ID of the rented book",
                                                "type": "integer"
                                            },
                                            "rented_at": {
                                                "description": "Date and time when the book was rented",
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "due_at": {
                                                "description": "Due date for returning the book",
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "returned_at": {
                                                "description": "Date and time when the book was returned",
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "overdue": {
                                                "description": "Indicates if the rental is overdue",
                                                "type": "boolean"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "User ID is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No rental history found for this user."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}